pull request- TC-35 updated code for  corrections made by Code review
