+++ src/main/java/org/umssdiplo/automationv01/core/managepage/accident/AccidentList.java
@@ -18,8 +18,4 @@ public class AccidentList extends BasePage {
         return CommonEvents.isPresent(btnCreateAccident);
     }
 
-    public void clickOnCreateAccidentButton() {
-        CommonEvents.clickButton(btnCreateAccident);
-    }
-
 }
