+++ src/main/java/org/umssdiplo/automationv01/core/managepage/home/SSIAHome.java
@@ -3,8 +3,8 @@ package org.umssdiplo.automationv01.core.managepage.home;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.umssdiplo.automationv01.core.managepage.BasePage;
-import org.umssdiplo.automationv01.core.managepage.accident.AccidentList;
 import org.umssdiplo.automationv01.core.managepage.employee.EmployeeList;
+import org.umssdiplo.automationv01.core.managepage.menuheader.safetyMenu.SafetyMenu;
 import org.umssdiplo.automationv01.core.managepage.role.RoleList;
 import org.umssdiplo.automationv01.core.utils.CommonEvents;
 
@@ -27,14 +27,11 @@ public class SSIAHome extends BasePage {
     @FindBy(id = "safetyHealthAdministrationId")
     private WebElement safetyMenu;
 
-    @FindBy(id = "accidentMenu")
-    private WebElement accidentMenu;
-
-    public void clickOnPersonnelMenu(){
+    public void clickOnPersonnelMenu() {
         CommonEvents.clickButton(personnelMenu);
     }
 
-    public EmployeeList clickOnEmployeeMenu(){
+    public EmployeeList clickOnEmployeeMenu() {
         CommonEvents.clickButton(employeeMenu);
         return new EmployeeList();
     }
@@ -44,12 +41,9 @@ public class SSIAHome extends BasePage {
         return new RoleList();
     }
 
-    public void clickOnSafetyMenu() {
+    public SafetyMenu clickOnSafetyMenu() {
         CommonEvents.clickButton(safetyMenu);
+        return new SafetyMenu();
     }
 
-    public AccidentList clickOnAccidentMenu() {
-        CommonEvents.clickButton(accidentMenu);
-        return new AccidentList();
-    }
 }
